// src/models/elements-form.ts
export interface FormData {
  number: number
  name: string
  atomic_mass: number
  appearance: string
  boil: number
  category: string
  density: number
  melt: number
  molar_heat: number
  discovered_by: string | null
  named_by: string | null
  period: number
  phase: string
  symbol: string
  source: string | null
  spectral_img: string | null
  summary: string | null
  electron_configuration: string
  electron_configuration_semantic: string
  electron_affinity: number
  element_img: string
  xpos: number
  ypos: number
  enabled: boolean
}